[CleanCode] - Refactor microsservice routes
