+++ src/index.js
@@ -1,7 +1,7 @@
 const express = require('express');
 const mongoose = require('mongoose');
 const cors = require('cors');
-const routes = require('./routes');
+const routes = require('../src/Routes/routes');
 require('dotenv').config();
 
 const {
